requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
flask-mysql-connector==1.1.0
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.23.1
pandas==1.4.3
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
Werkzeug==2.2.2
